Update due to "avar" comments.
(00:33:51) avar: SethRobertson: The reason I rewrite my history is two-fold: 1) To ensure that each commit works, sometimes this is not the case with commits I make locally (they may not even compile) 2) Amend / reorder / squash them so they can be clearly understood by me / others later on.
(00:34:03) avar: wrt the "Commit early and often" section
(00:34:37) avar: 1) is very important for bisecting, and 2) is for archeology / understanding why something is the way it is.
(00:38:27) avar: "your work will not be lost for at least two weeks unless you really work at it."
(00:38:43) avar: That should be "unless you manually purge it" or something